requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

jax[cuda12_pip] 
flax
ott-jax[neural]
tensorflow==2.15.1
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-probability==0.23.0
wandb
absl-py
Pillow
ml-collections
mvlearn
dm-pix
ipykernel
datasets
umap-learn
